calendar rendering troubles
